+++ CHANGELOG.md
@@ -40,12 +40,14 @@ All notable changes to this project will be documented in this file.
 - check all TinyDB
 
 ### Improvements
+- Add `normalize_issn` in `triplea.service.repository.export.dataset` version 1.1.1.003
 - Add more field in affilation_integration in dataset(`_json_converter_03`) version 1.1.1.002
 - Add `_get_affiliation_list_from_all_bank` and `_affiliation_mining_multiple_parser_in_list` and define AffiliationParseMethod.REGEX_API 2025-08-04
 - Integrated all task improvment
 - In version 0.0.7 the datamodel change. we change version to major
 
 ### Bug Fixes
+- Bug fix in `clean_publication_type` version 1.1.1.003
 
 ## v0.0.7 2025-07-06